Push node updates to ui

+++ src/decode/model/Decoder.cpp
@@ -193,7 +193,7 @@ StatusDecoder::~StatusDecoder()
 {
 }
 
-bool StatusDecoder::decode(uint8_t msgId, bmcl::Bytes payload)
+bool StatusDecoder::decode(uint64_t msgId, bmcl::Bytes payload)
 {
     auto it = _decoders.find(msgId);
     if (it == _decoders.end()) {
